requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
playwright==1.40.0
moviepy==1.0.3
python-ffmpeg==2.0.12
edge-tts==6.1.9
openai==1.3.7
google-api-python-client==2.109.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
sqlalchemy[aiosqlite]==2.0.23
alembic==1.12.1
apscheduler==3.10.4
loguru==0.7.2
python-dotenv==1.0.0
typer[all]==0.9.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0
Pillow==10.1.0
elevenlabs==0.2.26
python-multipart==0.0.6
loguru>=0.7.0